Update editor.js
Lines 182-184: The only time the main.py file tab will be added is when the webpage is loading; since the user will be on main.py when the webpage loads we’ll make it the current file.
Lines 229-230: Remove the class “current” from the current file and add the class “current” to the new current file. (Note: On line 229 I’ve used .editorTab.current not .current so if the class “current” is used later it won’t interfere with it.)
Haven’t tested it yet; will before submitting a pull request.