Clean up some dependencies in js files
This was an attempt to make it a bit clearer what the dependencies were
between the JavaScript files used for displaying and running a sketch in
PyAngelo. There is now an EditorSetup.js file that creates a class for
an Ace Editor. There is also a SkulptSetup.js file that has the basic
requirements for running Skulpt on a webpage. Any page that uses Skulpt
can include these 2 files and then create any other functionality as
required.