Partial auto save implemented
Whenever the editor is changed, the code will check if 10 seconds has
passed since the last save and if so will save the current file again.
This means that there will be a maximum of 10 seconds worth of work that
has not been saved. The start and stop buttons still should be used as
the main way of saving work.